HTML, CSS with Webpack
styling many pages, especially videoplayer like Youtube.

+++ src/controllers/videoController.js
@@ -1,37 +1,52 @@
 import Video from "../models/Video";
-
+import User from "../models/User";
 
 
 export const home = async(req, res) => {
-  const videos = await Video.find({}).sort({ createdAt: "asc"})
-  return res.render("home", { pageTitle: "Home", videos});
+  const videos = await Video.find({})
+    .sort({ createdAt: "desc"})
+    .populate("owner");
+  return res.render("root/home", { pageTitle: "Home", videos});
 };
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
-  if(!video) {
+  const video = await Video.findById(id).populate("owner");
+  if (!video) {
     return res.render("404", { pageTitle: "Video not found."});
   }
-  return res.render("watch", { pageTitle: video.title, video });
+  return res.render("videos/watch", { pageTitle: video.title, video });
 };
 
 
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
-  if(!video) {
+  if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found."});
   }
-  return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
+  if (String(video.owner) !== String(_id)) { // video.owner = object, _id = string
+    return res.status(403).redirect("/");
+  }
+  return res.render("videos/edit", { pageTitle: `Edit: ${video.title}`, video });
 };
+
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({_id: id});
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
   if(!video) {
     return res.render("404", { pageTitle: "Video not found."});
+  }
+  if(String(video.owner) !== String(_id)) { // video.owner = object, _id = string
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id,{
     title,
@@ -44,16 +59,24 @@ export const postEdit = async (req, res) => {
 
 
 export const getUpload = (req, res) => {
-  return res.render("upload", { pageTitle: "Upload Video" });
+  return res.render("videos/upload", { pageTitle: "Upload Video" });
 };
+
 export const postUpload = async(req, res) => {
+  const { user: { _id }, } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -67,7 +90,21 @@ export const postUpload = async(req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.render("404", { pageTitle: "Video not found."});
+  }
+  if (String(video.owner) !== String(_id)) { // video.owner = object, _id = string
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id),1);
+  user.save();
   return res.redirect("/");
 };
 
@@ -81,7 +118,7 @@ export const search = async (req, res) => {
       title: {
         $regex: new RegExp(`${keyword}$`, "i"),
       },
-    });
+    }).populate("owner");
   }
-  return res.render("search", { pageTitle: "Search", videos });
+  return res.render("root/search", { pageTitle: "Search", videos });
 };
